==> P0_chords_storage.ipynb <==
import pandas as pd

octave_notes = ['C', 'C#', 'D', 'D#', 'E', 'F', 'F#', 'G', 'G#', 'A', 'A#', 'B']

def transpose_to_c(song, scale):
    notes = song.split(' ')
    new_scale = 'C'
    diff = 12 - octave_notes.index(scale)
    new_notes = []
    for i in notes:
        x = (octave_notes.index(i) + diff)%12
        new_notes.append(octave_notes[x])
    return new_notes

chords = [
    [['G C G G C D C G C D C D C D G'], ['G C C D D G G C C D D D D G C G'], ['G C G D C D G C D G'], 'Waltz', 'G'],
    [['G G D D C C C C G G D D C C G G D D'], ['G G G C C D D G G G G G C C D D G G C C D D G'], ['G C D C G G C D C G'], 'Waltz', 'G'],
    [['C G C G C F G C C F C F C F C G'], ['C G C G C G F C C F C G'], ['C F G C C F G C'], 'Classic', 'C'],
    [['C F C F C C F C F C C F G C C F G C'], ['C F G C C F G C C F G C F G C'], ['C F C F G F G C C F G G C'],'Classic' , 'C'],
    [['G C G D C D G C D C D G'], ['G G C D G G C D G G D G C D G'], ['G C D C G C D C G'],'Classic' , 'G'],
    [['C C F G C F G C F G C F G C'], ['C G F G C G C G F G C F F G F G C'], ['C F G F C F G F G C'], 'Waltz' , 'C'],
    [['C F G F C C F G F G F C C F G F G C'], ['C C F F C C C G F G C C G F G C'], ['C F G F C G F G F C'],'Classic' , 'C'],
    [['E B A B A B E A B E A B E'], ['E A E B A E E B A E E B A E'], ['E A B E A B A E'],'Waltz' ,'E'],
    [['D G A D D G A D D A G A G D A'], ['D D A G A D D'], ['D G D A D G D A G D'],'Classic' ,'D'],
    [['C C C F G C F C G'], ['C F C G C F C G'], ['C C F F C F G C'],'Classic' ,'C'],
    [['C C F C C F G G F F G G F G C'], ['C C G G F G C C C F G'], ['C F C G F C F G F C'],'Classic' ,'C'],
    [['G C G D D C D C G C D C G'], ['G D C D G G D C D G G C G C D C G'], ['G C C D D C C G'],'Waltz' ,'G'],
    [['G D C D C D G G D C D C G'], ['G G D D C D G'], ['G C D D C D G'],'Classic' ,'G'],
    [['C C F F G F C C G F C F G F C'], ['C G F C F G C'], ['C G F G C G F G C F G C'], 'Waltz','C'],
    [['D D G A G D D A G A D'], ['D A G D A D D G D A D'], ['D D D A G D D D D A G D'],'Waltz', 'D'],
    [['G D G G D G C D C D C D G'], ['G C G G C G C D C D G C G'], ['G C G C D C G D G C G'], 'Waltz', 'G'],
    [['C F G C G F C F G F G'], ['C F C G F G C F C G C'], ['C F C F G C F C'], 'Classic', 'C'],
    [['D G A G A D D G A G D'], ['D A G A D D A G A'], ['D G D G A A G D'], 'Classic', 'D'],
    [['C G C G F G C C C F G F C'], ['C C F G C C F G'], ['C G G F F G F C'], 'Classic', 'C'],
    [['E E A A B B A B A E'], ['E A E E B E E A B A E'], ['E A B A B E A B E'], 'Waltz', 'E']
]

chords_transposed = []
for i in range(len(chords)):
    temp = []
    temp.append(transpose_to_c(chords[i][0][0], chords[i][4]))
    temp.append(transpose_to_c(chords[i][1][0], chords[i][4]))
    temp.append(transpose_to_c(chords[i][2][0], chords[i][4]))
    temp.append(chords[i][3])
    temp.append('C')
    chords_transposed.append(temp)

print(chords_transposed)

df = pd.DataFrame(chords_transposed ,columns =['Stanza', 'Chorus', 'Interlude', 'Type', 'Scale'])
df

df.to_csv('data.csv')
