==> P2_find_chords.ipynb <==
### Finding the chord families of each scale

#chord_family = { 'C' : ['C', 'F', 'G', 'Am', 'Dm', 'Em']}
chord_family = { 'C' : ['C', 'F', 'G']}
octave_notes = ['C', 'C#', 'D', 'D#', 'E', 'F', 'F#', 'G', 'G#', 'A', 'A#', 'B']
for i in range(1, len(octave_notes)):
    x = octave_notes[i]
    l = []
    for j in chord_family['C']:
        j2 = j[0]
        temp = octave_notes.index(j2)
        temp2 = (i + temp)%(len(octave_notes))
        new = octave_notes[temp2]
        if len(j) > 1:
            new += 'm'
        l.append(new)
    chord_family[octave_notes[i]] = l
    
chord_family

### Finding the notes in each chord

chord_notes_major = { 'C' : ['C', 'E', 'G']}
for i in range(1, len(octave_notes)):
    x = octave_notes[i]
    l = []
    for j in chord_notes_major['C']:
        temp = octave_notes.index(j)
        temp2 = (i + temp)%(len(octave_notes))
        new = octave_notes[temp2]
        l.append(new)
    chord_notes_major[octave_notes[i]] = l
    
chord_notes_major

chord_notes_minor = { 'Cm' : ['C', 'D#', 'G']}
for i in range(1, len(octave_notes)):
    x = octave_notes[i]
    l = []
    for j in chord_notes_minor['Cm']:
        temp = octave_notes.index(j)
        temp2 = (i + temp)%(len(octave_notes))
        new = octave_notes[temp2]
        l.append(new)
    chord_notes_minor[octave_notes[i] + 'm'] = l
    
chord_notes_minor

### Find chords from notes

def remove_octave(notes):
    for i in range(len(notes)):
        notes[i] = notes[i][:-1]
    return notes

#notex = ['C0', 'C4', 'C#4', 'C#4', 'B4', 'C4', 'C5', 'C5', 'A3', 'F3', 'F3', 'D4', 'D4', 'F2', 'F2', 'F2', 'F2', 'C5', 'F2', 'F2', 'F3', 'F2', 'F2', 'F4', 'F2', 'F2', 'C4', 'F3', 'F2', 'F2', 'G#2', 'G2', 'G#2', 'G2', 'G#2', 'D4', 'G2', 'G2', 'G2', 'G#2', 'G2', 'G#2', 'G2', 'G2', 'G2', 'G2', 'G2', 'G2', 'G#2', 'G2', 'C#4', 'G#2', 'G2', 'G2', 'G2', 'G2', 'G2', 'G#2', 'G2', 'G2', 'G2', 'G2', 'G2', 'G2', 'G#2', 'G2', 'G4', 'G2', 'G2', 'G#2', 'G2', 'G#2', 'G2', 'D#2', 'F3', 'F3', 'G#2', 'F2', 'F2', 'F2', 'F2', 'F2', 'F3', 'F2', 'F2', 'F#3', 'F3', 'F2', 'C#4', 'F2', 'C#4', 'C#4', 'C#4', 'C4', 'F2', 'C#5', 'C5', 'F3', 'C5', 'F2', 'F2', 'F3', 'F3', 'F2', 'A4', 'F2', 'F2', 'F2', 'F3', 'F3', 'F#3', 'F2', 'F3', 'F2', 'F#4', 'F3', 'F2']
notex = ['G4', 'G4', 'G4', 'B4', 'A4', 'G4', 'G4', 'G4', 'D5', 'C5', 'B4', 'E4', 'D4', 'C4', 'B4', 'A4', 'B4', 'F#4', 'G4', 'G4', 'G4', 'B4', 'A4', 'G4', 'D5', 'C5', 'B4', 'E5', 'D5', 'C5', 'B4', 'A4', 'B4', 'F#4', 'G4', 'F#4', 'E4', 'F#4', 'G4', 'F#4', 'B4', 'D5', 'B4', 'D5', 'E5', 'D5', 'A4', 'G4', 'B4', 'D5', 'B4', 'D5', 'E5', 'D5', 'A4', 'G4', 'G4', 'A4', 'C5', 'E5', 'E5', 'G4', 'A4', 'G4', 'A4', 'C4', 'B4', 'D4']
notex = remove_octave(notex)
notes = []
for i in notex:
    if len(i) == 2:
        notes.append(i[0])
    else:
        notes.append(i[0:2])

# print(notes)
scale = 'G'

def find_chord(note, scale):
    fam = chord_family[scale]
    family_notes = []
    for i in fam:
        if i[-1] == 'm':
            family_notes.append(chord_notes_minor[i])
        else:
            family_notes.append(chord_notes_major[i])
    simi = []
    posi = []
    for i in range(len(fam)):
        if note in family_notes[i]:
            if note == family_notes[i][0]:
                simi.append(101)
                posi.append(0)
            else:
                simi.append(100)
                posi.append(family_notes[i].index(note))
        else:
            simi.append(0)
            posi.append(-1)
    if max(simi) == 0:
        return 'a', -1
    x = simi.index(max(simi))
    chord = fam[x]
    pos = posi[x]
    return chord, pos

#beat = input('Enter beat type : Classic, Waltz, 6/8 - ').capitalize()

def find_chord_prog(notes, scale, beat):  
    i = 0
    chord_prog = []
    pos_prog = []
    if beat == 'Waltz':
        a = 3
    elif beat == 'Classic':
        a = 4
    elif beat == '6/8':
        a = 6
    else :
        a = 4
    while i < len(notes):
        note = notes[i]
        chord, pos = find_chord(note, scale)
        if chord == 'a':
            chord_prog.append(chord_prog[-1])
            pos_prog.append(pos_prog[-1])
        else:
            chord_prog.append(chord)
            pos_prog.append(pos)
        i += a
    return chord_prog, pos_prog
    
#chord_prog, pos_prog = find_chord_prog(notes, scale, beat)

# print(chord_prog)
# print(pos_prog)

def transpose_chord(prog, scale, new_scale):
    octave_notes = ['C', 'C#', 'D', 'D#', 'E', 'F', 'F#', 'G', 'G#', 'A', 'A#', 'B']
    diff = (octave_notes.index(new_scale) - octave_notes.index(scale))%12
    new_notes = []
    for i in prog:
        x = (octave_notes.index(i) + diff)%12
        new_notes.append(octave_notes[x])
    return new_notes
