==> P3_music_generation.ipynb <==
#pip install ipynb

import pandas as pd
import ast
import random
#from ipynb.fs.full.Scale_identification_and_transposing import identify_scale, transpose

def find_max(mat):
    x = []
    for i in range(len(mat)):
        x.append(max(mat[i]))
    return max(x)

def GET_SCORE(n1, n2, penalty = -1, reward = 1):
     
    if n1 == n2:
        return reward
    else:
        return penalty

import numpy as np
  
def alignment(X, Y, penalty = -1, reward = 1):
    # initialize score matrix
    score_matrix = np.ndarray((len(X) + 1, len(Y) + 1))
    for i in range(len(X) + 1):
        score_matrix[i, 0] = penalty * i
    for j in range(len(Y) + 1):
        score_matrix[0, j] = penalty * j
    # define each cell in the matrix by as the max score possible in that stage
    for i in range(1, len(X) + 1):
        for j in range(1, len(Y) + 1):
            match = score_matrix[i - 1, j - 1] + GET_SCORE(X[i - 1], Y[j - 1], penalty, reward)
            delete = score_matrix[i -1, j] + penalty
            insert = score_matrix[i, j - 1] + penalty    
            score_matrix[i, j] = max([match, delete, insert])
             
    i = len(X)
    j = len(Y)
    align_X = ""
    align_Y = ""
    maxc = find_max(score_matrix)
     
#     while i > 0 or j > 0:
#         current_score = score_matrix[i, j]
#         left_score = score_matrix[i - 1, j]
#         if i > 0 and j > 0 and X[i - 1] == Y[j - 1]:
#             align_X = X[i - 1] + align_X
#             align_Y = Y[j - 1] + align_Y
#             i = i - 1
#             j = j - 1
#         elif i > 0 and current_score == left_score + penalty:
#             align_X = X[i - 1] + align_X
#             align_Y = "-" + align_Y
#             i = i - 1   
#         else:
#             align_X = "-" + align_X
#             align_Y = Y[j - 1] + align_Y
#             j = j - 1
            
    return maxc

a = ['C', 'F', 'C', 'C', 'F', 'G', 'F', 'C', 'F', 'G', 'F', 'G', 'F', 'G', 'C']
b = ['C', 'C', 'F', 'G', 'C', 'F', 'G', 'C', 'F', 'G', 'C', 'F', 'G', 'C']
#print(alignment(a,b))

df = pd.read_csv('data.csv')
df

# song_type = input('Enter song type Classic / Waltz : ').capitalize()
# chord_stanza = input('Enter stanza chord sequence without any space : ').upper()
# chord_chorus = input('Enter chorus chord sequence without any space : ').upper()

def music_gen(df, song_type, chord_stanza, chord_chorus):
    df2 = df[df['Type'] == song_type].reset_index(drop = True)
    df_stanza = df2['Stanza']
    df_chorus = df2['Chorus']
    score = []
    for i in range(len(df2)):
        x = ast.literal_eval(df_stanza[i])
        y = ast.literal_eval(df_chorus[i])
        xs = alignment(chord_stanza, x)
        ys = alignment(chord_chorus, y)
        score.append((2*xs)+(3*ys))
    bestmatch = df2.loc[[score.index(max(score))]]
    x = bestmatch['Interlude'].to_list()
    x = ast.literal_eval(x[0])
    return score, x, bestmatch
    
# score, x, bestmatch = music_gen(df, song_type, chord_stanza, chord_chorus)
# print(score)
# print(max(score))
# print('Stanza = ' + bestmatch['Stanza'])
# print(f'Chorus = ' + bestmatch['Chorus'])
# print(f'Interlude = ' + bestmatch['Interlude'])
# print(x)

seq = [0, 1, 0, 0, 0, 2, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 2, 2, 1, 0, 1]

def generate_op(seq, x, song_type):    
    l = min(len(x), len(seq))
    chord_op = {'C' : ['C', 'F', 'G'], 'F' : ['F', 'A', 'C'], 'G' : ['G', 'B', 'D']}
    scale_seq = ['C', 'D', 'E', 'F', 'G', 'A', 'B']
    note_op = []
    for i in range(l):
        note_op.append(chord_op[x[i]][seq[i]])
    i = 0
    note2 = []
    while i < len(note_op)-1:
        note2.append(note_op[i])
        aind = scale_seq.index(note_op[i])
        bind = scale_seq.index(note_op[i+1])
        mi = min(aind, bind)
        ma = max(aind, bind)
        if song_type == 'Waltz':
            a = 2
        else:
            a = 3
        if mi == ma:
            tempj = [-1, 2, 1]
            for j in range(a):
                temp = (mi + tempj[j])%len(scale_seq)
                note2.append(scale_seq[temp])
        else:    
            for j in range(a):
                temp = random.randint(mi, ma)
                note2.append(scale_seq[temp])
        i += 1
    note2.append(note_op[i])
    return note2

#print(generate_op(seq, x, song_type))
