==> P4_music_output.ipynb <==
#pip install musicalbeeps

#pip install scamp

# import musicalbeeps


# player = musicalbeeps.Player(volume = 0.3,
#                             mute_output = False)

notes = ['C4', 'A3', 'C4', 'F4', 'C4', 'G4', 'F4', 'F4', 'F4', 'F4', 'F4', 'A4', 'B4', 'F4', 'F4', 'F4', 'B4', 'C5']
# for i in range(len(notes)):
#     player.play_note(notes[i])

from scamp import *
import numpy as np
import random

# session = Session()
# piano1 = session.new_part("Piano")
# pitches = [64, 66, 71, 73, 74, 66, 64]
# for pitch in pitches:
#     piano1.play_note(pitch, 1.0, 0.25)

notes = ['C4', 'C4', 'D4', 'D4', 'E4', 'E4', 'C4', 'C4', 'D4', 'D4', 'E4', 'E4', 'C4', 'C4', 'D4', 'D4', 'E4', 'E4', 'C4', 'F4', 'A4', 'A4', 'G4', 'G4', 'C5', 'C5', 'A4', 'A4', 'F4', 'F4', 'C5', 'C5', 'B4', 'B4', 'G4', 'G4', 'C5', 'C5', 'A4', 'A4', 'F4', 'F4', 'D5', 'D5', 'C5', 'D5']

def play_piano(notes):
    name = ["C0","C#0","D0","D#0","E0","F0","F#0","G0","G#0","A0","A#0","B0",
            "C1","C#1","D1","D#1","E1","F1","F#1","G1","G#1","A1","A#1","B1",
            "C2","C#2","D2","D#2","E2","F2","F#2","G2","G#2","A2","A#2","B2",
            "C3","C#3","D3","D#3","E3","F3","F#3","G3","G#3","A3","A#3","B3",  
            "C4","C#4","D4","D#4","E4","F4","F#4","G4","G#4","A4","A#4","B4",
            "C5","C#5","D5","D#5","E5","F5","F#5","G5","G#5","A5","A#5","B5",
            "C6","C#6","D6","D#6","E6","F6","F#6","G6","G#6","A6","A#6","B6",
            "C7","C#7","D7","D#7","E7","F7","F#7","G7","G#7","A7","A#7","B7",
            "C8","C#8","D8","D#8","E8","F8","F#8","G8","G#8","A8","A#8","B8","C9+"]
    pitches = []
    for i in range(len(notes)):
        pitches.append(name.index(notes[i]))
    session = Session()
    piano1 = session.new_part("Piano")
    for pitch in pitches:
        piano1.play_note(pitch, 1.0, 0.5)

def marked_output(op):
    piano_notes = ["C0","C#0","D0","D#0","E0","F0","F#0","G0","G#0","A0","A#0","B0",
                   "C1","C#1","D1","D#1","E1","F1","F#1","G1","G#1","A1","A#1","B1",
                   "C2","C#2","D2","D#2","E2","F2","F#2","G2","G#2","A2","A#2","B2",
                   "C3","C#3","D3","D#3","E3","F3","F#3","G3","G#3","A3","A#3","B3",  
                   "C4","C#4","D4","D#4","E4","F4","F#4","G4","G#4","A4","A#4","B4",
                   "C5","C#5","D5","D#5","E5","F5","F#5","G5","G#5","A5","A#5","B5",
                   "C6","C#6","D6","D#6","E6","F6","F#6","G6","G#6","A6","A#6","B6",
                   "C7","C#7","D7","D#7","E7","F7","F#7","G7","G#7","A7","A#7","B7",
                   "C8","C#8","D8","D#8","E8","F8","F#8","G8","G#8","A8","A#8","B8","C9+"]
    marked_op = [op[0]+'5']
    for i in range(1, len(op)):
        x = int(marked_op[i-1][-1])
        temp = [op[i]+str(x-1), op[i]+str(x), op[i]+str(x+1)]
        ind = piano_notes.index(marked_op[i-1])
        temp_dis = []
        for j in range(3):
            temp_dis.append(abs(ind - piano_notes.index(temp[j])))
        mindis = temp_dis.index(min(temp_dis))
        marked_op.append(temp[mindis])
    return marked_op
    

op = ['C', 'A', 'C', 'C', 'F', 'D', 'F', 'F', 'F', 'B', 'A', 'B', 'A', 'B', 'F']
#marked_op = marked_output(op)
#print(marked_op)

# pitches = []
# for i in range(len(marked_op)):
#     pitches.append(name.index(marked_op[i]))
    
# session = Session()
# piano1 = session.new_part("Piano")

# for pitch in pitches:
#     piano1.play_note(pitch, 1.0, 0.5)

# def build_chord(interval_options,
#                 num_notes,
#                 pitch_center,
#                 round_transposition=True):
    
#     chord = [0]
#     while len(chord) < num_notes:
#         chord.append(chord[-1] + random.choice(interval_options))
    
#     transposition = pitch_center - sum(chord) / len(chord)
#     if round_transposition:
#         transposition = round(transposition)
#     return [p + transposition for p in chord]


# s = Session()

# organ = s.new_part("Acoustic Guitar")


# for i in range(5):
#     organ.play_chord(
#         build_chord([2, 5, 7], random.randint(2, 10), random.randint(40, 80)),
#         0.8,
#         0.5,
#         "staccato" if random.random() < 0.5 else None
#     )
